requirements.txt
fastapi
uvicorn
torch
transformers
sentence-transformers
nltk
spacy
faiss-cpu
python-dotenv
